Breaks code in smaller files && adjusts grunt tasks accordingly
