+++ src/common/gridster-debounce.factory.js
@@ -1,0 +1,26 @@
+(function(angular) {
+	'use strict';
+
+	angular.module('gridster').factory('gridsterDebounce', function() {
+		return function gridsterDebounce(func, wait, immediate) {
+			var timeout;
+			return function() {
+				var context = this,
+					args = arguments;
+				var later = function() {
+					timeout = null;
+					if (!immediate) {
+						func.apply(context, args);
+					}
+				};
+				var callNow = immediate && !timeout;
+				clearTimeout(timeout);
+				timeout = setTimeout(later, wait);
+				if (callNow) {
+					func.apply(context, args);
+				}
+			};
+		};
+	});
+
+})(window.angular);
