+++ src/components/gridster-no-drag/gridster-no-drag.directive.js
@@ -1,0 +1,12 @@
+(function(angular) {
+	'use strict';
+
+	angular.module('gridster').directive('gridsterNoDrag', function() {
+		return {
+			restrict: 'A',
+			link: function(scope, $element) {
+				$element.addClass('gridster-no-drag');
+			}
+		};
+	});
+})(window.angular);
